// src/types/errors.ts
export enum Errors {
	NO_API_KEY = "No API key provided",
	INVALID_API_KEY = "Invalid API key",
	USER_NOT_FOUND = "User not found",
	NO_LOGGER_WITH_NAME = "No logger with that name",
	NO_MESSAGE = "No message provided",
	NO_LEVEL = "No level provided",
	NO_LOGGER_NAME = "No logger name provided",
}